+++ komikaan.Data/Enums/DisruptionStage.cs
@@ -1,0 +1,8 @@
+﻿namespace komikaan.Data.Enums
+{
+    public enum DisruptionStage
+    {
+        Finished,
+        Ongoing
+    }
+}