+++ komikaan/Context/NSContext.cs
@@ -3,6 +3,7 @@ using komikaan.Enums;
 using komikaan.Interfaces;
 using komikaan.Models;
 using komikaan.Models.API.NS;
+using Refit;
 
 namespace komikaan.Context
 {
@@ -20,22 +21,29 @@ namespace komikaan.Context
         {
             _nsApi = nsApi;
             _logger = logger;
+            _allDisruptions = new List<SimpleDisruption>();
+            _allStations = new Dictionary<string, Station>();
         }
 
-        public async Task LoadRelevantData(CancellationToken cancellationToken)
+        public async Task StartAsync(CancellationToken cancellationToken)
         {
-            await GetAllData();
+            await GetAllDataAsync();
         }
 
-        public async Task<IEnumerable<SimpleDisruption>> GetDisruptions(string from, string to)
+        public async Task LoadRelevantDataAsync(CancellationToken cancellationToken)
+        {
+            await GetAllDataAsync();
+        }
+
+        public Task<IEnumerable<SimpleDisruption>> GetDisruptionsAsync(string from, string to)
         {
             var fromStop = _allStations[from];
             var toStop = _allStations[to];
 
-            var relevantDisruptions = _allDisruptions.Where(disruption => disruption.AffectedStops.Any(stop => stop.Equals(fromStop.code, StringComparison.InvariantCultureIgnoreCase) || stop.Equals(toStop.code, StringComparison.InvariantCultureIgnoreCase))).ToList();
-            relevantDisruptions = relevantDisruptions.FindAll(disruption => disruption.Active).ToList();
+            var relevantDisruptions = _allDisruptions.Where(disruption => disruption.AffectedStops.Any(stop => stop.Equals(fromStop.code, StringComparison.InvariantCultureIgnoreCase) || stop.Equals(toStop.code, StringComparison.InvariantCultureIgnoreCase)));
+            relevantDisruptions = relevantDisruptions.ToList().FindAll(disruption => disruption.Active);
 
-            return relevantDisruptions;
+            return Task.FromResult(relevantDisruptions);
         }
 
         public Task<IEnumerable<SimpleDisruption>> GetAllDisruptions(bool active)
@@ -49,7 +57,7 @@ namespace komikaan.Context
             return Task.FromResult(_allStations);
         }
 
-        public async Task<IEnumerable<SimpleTravelAdvice>> GetTravelAdvice(string from, string to)
+        public async Task<IEnumerable<SimpleTravelAdvice>> GetTravelAdviceAsync(string from, string to)
         {
             var fromStop = _allStations[from];
             var toStop = _allStations[to];
@@ -58,33 +66,43 @@ namespace komikaan.Context
             var simplifiedTravelAdvices = new List<SimpleTravelAdvice>();
             foreach (var trip in travelAdvice.trips)
             {
-                simplifiedTravelAdvices.Add(GenerateTravelAdvice(from, trip));
+                var simplifiedTravelAdvice = GenerateSimplifiedTravelAdvice(from, trip);
+                simplifiedTravelAdvices.Add(simplifiedTravelAdvice);
             }
             return simplifiedTravelAdvices;
         }
 
-        private SimpleTravelAdvice GenerateTravelAdvice(string from, Trip trip)
+        private SimpleTravelAdvice GenerateSimplifiedTravelAdvice(string from, Trip trip)
         {
             var simpleTravelAdvice = new SimpleTravelAdvice();
             simpleTravelAdvice.Source = DataSource.NederlandseSpoorwegen;
             simpleTravelAdvice.PlannedDurationInMinutes = trip.plannedDurationInMinutes;
             simpleTravelAdvice.ActualDurationInMinutes = trip.actualDurationInMinutes;
-            simpleTravelAdvice.Route = new List<SimpleRoutePart>();
-
-            //Add the first station manually
-            simpleTravelAdvice.Route.Add(new SimpleRoutePart()
+            simpleTravelAdvice.Route = new List<SimpleRoutePart>
             {
-                Cancelled = false,
-                Name = from,
-                RealisticTransfer = true,
-                PlannedArrival = null,
-                ActualArrival = null
-            });
+                //Add the first station manually
+                new SimpleRoutePart()
+                {
+                    Cancelled = false,
+                    Name = from,
+                    RealisticTransfer = true,
+                    PlannedArrival = null,
+                    ActualArrival = null
+                }
+            };
 
+            CalculateLegs(trip, simpleTravelAdvice);
+
+            return simpleTravelAdvice;
+        }
+
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Maintainability", "AV1500:Member or local function contains too many statements", Justification = "Collection of leg data, splitting this causes arguably more confusion in the current form")]
+        private static void CalculateLegs(Trip trip, SimpleTravelAdvice simpleTravelAdvice)
+        {
             // We see these objects different then the NS. We look at it per station as thats in our interest
             // While the NS looks at it per leg, so sometimes we have to look at the previous item to look forward
             var previous = simpleTravelAdvice.Route.First();
-            Leg previousLeg = null;
+            Leg? previousLeg = null;
 
             foreach (var leg in trip.legs)
             {
@@ -128,31 +146,37 @@ namespace komikaan.Context
 
                 previous.RealisticTransfer = true;
             }
-
-
-            return simpleTravelAdvice;
         }
 
-        private async Task GetAllData()
+        private async Task GetAllDataAsync()
         {
-            _allStations = await LoadAllStops();
-            _allDisruptions = await LoadAllDisruptions();
+            try
+            {
+                _allStations = await LoadAllStopsAsync();
+                _allDisruptions = await LoadAllDisruptionsAsync();
+            }
+            catch (ApiException apiException)
+            {
+                // A backoff should be implemented, for example Polly
+                _logger.LogError(apiException, "Failed to reload information");
+            }
         }
 
-        private async Task<List<SimpleDisruption>> LoadAllDisruptions()
+        private async Task<List<SimpleDisruption>> LoadAllDisruptionsAsync()
         {
             var rawDisruptions = await _nsApi.GetAllDisruptions();
 
             var data = new List<SimpleDisruption>();
             foreach (var disruption in rawDisruptions)
             {
-                data.Add(GenerateSimplifiedDisruption(disruption));
+                var simpleDisruption = GenerateSimplifiedDisruption(disruption);
+                data.Add(simpleDisruption);
             }
 
             return data;
         }
 
-        private async Task<Dictionary<string, Station>> LoadAllStops()
+        private async Task<Dictionary<string, Station>> LoadAllStopsAsync()
         {
             var stationRoot = await _nsApi.GetAllStations();
             var dict = new Dictionary<string, Station>();
@@ -164,30 +188,30 @@ namespace komikaan.Context
             return dict;
         }
 
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Maintainability", "AV1500:Member or local function contains too many statements", Justification = "Creation of simplified disruptions, splitting this causes arguably more confusion in the current form")]
         private SimpleDisruption GenerateSimplifiedDisruption(Disruption disruption)
         {
             _logger.LogInformation("Loading: {name}, active {act}", disruption.title, disruption.isActive);
-            var simpleDisruption = new SimpleDisruption();
-            simpleDisruption.Title = disruption.title;
-            var simpleText = disruption.timespans.Select(x => x.situation);
-            simpleDisruption.Descriptions = simpleText.Select(situation => situation.label);
-            var advices = disruption.timespans.Select(x => x.advices);
 
-            simpleDisruption.Advices = advices.SelectMany(advice => advice);
-            simpleDisruption.ExpectedEnd = disruption.end;
-
-            simpleDisruption.Active = disruption.isActive;
-
+            var situation = disruption.timespans.Select(timespan => timespan.situation);
+            var advices = disruption.timespans.Select(timespan => timespan.advices);
+            var affectedStops = disruption.publicationSections.SelectMany(section => section.section.stations.Select(disruptionStation => disruptionStation.stationCode)).ToList();
+            var description = situation.Select(situation => situation.label);
             if (!Enum.TryParse(disruption.type, true, out DisruptionType disruptionType))
             {
                 //Incase new disruption types are introduced
                 disruptionType = DisruptionType.Unknown;
             }
 
-            simpleDisruption.AffectedStops = disruption.publicationSections.SelectMany(section => section.section.stations.Select(disruptionStation => disruptionStation.stationCode)).ToList();
-
+            var simpleDisruption = new SimpleDisruption();
+            simpleDisruption.Title = disruption.title;
+            simpleDisruption.Descriptions = description;
+            simpleDisruption.Advices = advices.SelectMany(advice => advice);
+            simpleDisruption.ExpectedEnd = disruption.end;
+            simpleDisruption.Start = disruption.start;
+            simpleDisruption.Active = disruption.isActive;
+            simpleDisruption.AffectedStops = affectedStops;
             simpleDisruption.Source = Supplier;
-
             simpleDisruption.Type = disruptionType;
             return simpleDisruption;
         }