+++ komikaan/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.4.0] - 02-03-2024
+### Added
+- Catch API Exceptions from NS
+- Added StartDate to SimpleDisruption
+- DataSuppliers now have a real StartAsync again
+
 [1.3.0] - 29-02-2024
 ### Added
 - Add track data